+++ src/rex.forms/static/js/lib/widgets/readOnlyMatrix.js
@@ -10,6 +10,7 @@ var LabelRenderingMixin     = require('./LabelRenderingMixin');
 var SelfErrorRenderingMixin = require('./SelfErrorRenderingMixin');
 var matrixHeaderRow         = require('./matrixHeaderRow');
 var matrixRow               = require('./matrixRow');
+var FormEventsMixin         = require('./../form/FormEventsMixin');
 
 /**
  * Matrix
@@ -18,11 +19,13 @@ var readOnlyMatrix = React.createClass({
   mixins: [
     ReactForms.FieldsetMixin,
     LabelRenderingMixin,
-    SelfErrorRenderingMixin
+    SelfErrorRenderingMixin,
+    FormEventsMixin
   ],
 
   render: function() {
     var error = this.renderError();
+
     var className = cx(
       'rex-forms-Widget',
       'rex-forms-matrix',
@@ -31,7 +34,22 @@ var readOnlyMatrix = React.createClass({
       'rex-forms-readOnlyMatrix-' + this.props.name,
       error ? 'has-error' : null
     );
-    var questions = this.props.options.questions;
+
+    var events = this.formEvents();
+
+    var questions = this.props.options.questions.filter((question) => {
+      var rows = this.props.options.rows;
+
+      for (var r = 0; r < rows.length; r++) {
+        var val = this.value().get(rows[r].id);
+        if (!events.isHidden(question.fieldId, val)) {
+          return true;
+        }
+      }
+
+      return false;
+    });
+
     var rows = this.props.options.rows.map((row) =>
       <matrixRow
         readOnly={true}
@@ -41,6 +59,7 @@ var readOnlyMatrix = React.createClass({
         questions={questions}
         />
     );
+
     return (
       <div className={className}>
         {this.renderLabel()}